% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManip.R
\name{aggNext}
\alias{aggNext}
\title{Executes Aggregate functions on the next N records, similar to SQL Window functions}
\usage{
aggNext(vec, by = 1, aggfun, ignoreNA = TRUE)
}
\arguments{
\item{vec}{vector to aggregate}

\item{by}{number of next records to aggregate}

\item{aggfun}{the aggregate function. max, min, mean are all possible values}

\item{ignoreNA}{optional boolean. True by default. Set to FALSE if you do NOT want results when the window is less than "by" at the edges of your vector}
}
\value{
a vector containing the leading aggregate
}
\description{
Executes Aggregate functions on the next N records, similar to SQL Window functions
}
\examples{
y = c(5,7,2,9,3,8,1,7,4)
aggNext(y,5,mean)
}
