% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManip.R
\name{aggPrev}
\alias{aggPrev}
\title{Executes Aggregate functions on the previous N records, similar to SQL Window functions}
\usage{
aggPrev(vec, by = 1, aggfun, ignoreNA = TRUE, includeCurrent = FALSE)
}
\arguments{
\item{vec}{vector to aggregate}

\item{by}{number of previous records to aggregate}

\item{aggfun}{the aggregate function. max, min, mean are all possible values}

\item{ignoreNA}{optional boolean. True by default. Set to FALSE if you do NOT want results when the window is less than "by" at the edges of your vector}

\item{includeCurrent}{optional boolean, FALSE by default. Set to TRUE if you want to include the current row in the calculations}
}
\value{
a vector containing the lagged aggregate
}
\description{
Executes Aggregate functions on the previous N records, similar to SQL Window functions
}
\examples{
y = c(5,7,2,9,3,8,1,7,4)
aggPrev(y,5,mean)
}
