% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPlots.R
\name{drawPlots}
\alias{drawPlots}
\title{Draws the plots contained in a vector}
\usage{
drawPlots(plots, gridCols = 2)
}
\arguments{
\item{plots}{vector containing plots (must be ggplot2 objects)}

\item{gridCols}{optional, number of plots to be shown side-by-side}
}
\value{
nothing, it just plots
}
\description{
Utility function to draw plots on a grid
}
\examples{
plots = c(
    makeRegressionPlots(obs = simulated$y,pred = predict(rfTune, simulated),title = "RandomForests"),
    makeRegressionPlots(obs =  = test$y,pred = predict(rfTune, test),title = "RandomForests", trainSet = FALSE)
)
}
