% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPlots.R
\name{makeRegressionPlots}
\alias{makeRegressionPlots}
\title{Creates regression plots (Pred vs Obs and Pred vs Res)}
\usage{
makeRegressionPlots(obs, pred, title, trainSet = TRUE)
}
\arguments{
\item{obs}{vector containing **observed** response values}

\item{pred}{vector containing **predicted** response values}

\item{title}{title to show on the plot}

\item{trainSet}{optional boolean. True by default. Set to FALSE if you are plotting a test result. This controls the color-coding}
}
\value{
list of grobs containing the two plots
}
\description{
Utility function to create the plots with in-built color coding (so you won't confuse test and train results)
}
\examples{
plots = makeRegressionPlots(y, yhat, "Random Forests", TRUE)#'
}
