% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManip.R
\name{rollAgg}
\alias{rollAgg}
\title{Generic function to perform leading or lagged Aggregate functions}
\usage{
rollAgg(vec, by = 1, aggfun, ignoreNA = TRUE, roller = dplyr::lag,
  includeCurrent = FALSE)
}
\arguments{
\item{vec}{vector to aggregate}

\item{by}{number of previous records to aggregate}

\item{aggfun}{the aggregate function. max, min, mean are all possible values}

\item{ignoreNA}{optional boolean. True by default. Set to FALSE if you do NOT want results when the window is less than "by" at the edges of your vector}

\item{roller}{A function that produces the lagged or leading value. Set to dplyr::lag by default}

\item{includeCurrent}{optional boolean, FALSE by default. Set to TRUE if you want to include the current row in the calculations}
}
\value{
a vector containing the lagged aggregate
}
\description{
Generic function to perform leading or lagged Aggregate functions
}
\examples{
y = c(5,7,2,9,3,8,1,7,4)
aggPrev(y,5,mean)
}
